# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks/{taskId}/attachments:
    get:
      operationId: listAttachments
      tags: [attachments]
      summary: List attachments for a task
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: Attachments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attachment"
    post:
      operationId: registerAttachment
      tags: [attachments]
      summary: Register an uploaded file as a task attachment
      parameters:
        - $ref: "#/components/parameters/TaskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterAttachmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attachment"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /attachments/{attachmentId}:
    delete:
      operationId: deleteAttachment
      tags: [attachments]
      summary: Delete an attachment
      parameters:
        - $ref: "#/components/parameters/AttachmentId"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  parameters:
...
    TaskId:
      name: taskId
      in: path
      required: true
      schema:
        type: integer
...
    AttachmentId:
      name: attachmentId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Attachment:
      type: object
      required: [id, taskId, name, contentType, size, uploaderName, createdAt]
      properties:
        id:
          type: integer
        taskId:
          type: integer
        name:
          type: string
        contentType:
          type: string
        size:
          type: integer
        uploaderName:
          type: [string, "null"]
        createdAt:
          type: string
    RegisterAttachmentInput:
      type: object
      required: [name, contentType, size, objectPath]
      properties:
        name:
          type: string
          minLength: 1
        contentType:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        objectPath:
          type: string
          minLength: 1
...